feat: uint cast

+++ src/ZLinq.Unity/Assets/ZLinq.Unity/Runtime/NativeArrayExtensions.cs
@@ -74,7 +74,7 @@ namespace ZLinq.Linq
 
         public bool TryGetNext(out T current)
         {
-            if (index < source.Length)
+            if ((uint)index < (uint)source.Length)
             {
                 current = source[index++];
                 return true;
@@ -134,7 +134,7 @@ namespace ZLinq.Linq
 
         public bool TryGetNext(out T current)
         {
-            if (index < source.Length)
+            if ((uint)index < (uint)source.Length)
             {
                 current = source[index++];
                 return true;
